# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Los Banos Doctors Hospital HR API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      parameters:
        - in: query
          name: employeeId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeaveRequest" }
    post:
      operationId: createLeave
      tags: [leaves]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LeaveInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveRequest" }
  /leaves/{id}/advance:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    post:
      operationId: advanceLeave
      tags: [leaves]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalAction" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveRequest" }
  /leaves/balances/{employeeId}:
    parameters:
      - in: path
        name: employeeId
        required: true
        schema: { type: integer }
    get:
      operationId: getLeaveBalance
      tags: [leaves]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaveBalance" }
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      parameters:
        - in: query
          name: employeeId
          schema: { type: integer }
        - in: query
          name: type
          schema: { $ref: "#/components/schemas/RequestType" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/HrRequest" }
    post:
      operationId: createRequest
      tags: [requests]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HrRequestInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HrRequest" }
  /requests/{id}/advance:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    post:
      operationId: advanceRequest
      tags: [requests]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApprovalAction" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HrRequest" }
...
  /offboardings:
    get:
      operationId: listOffboardings
      tags: [offboarding]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Offboarding" }
    post:
      operationId: createOffboarding
      tags: [offboarding]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OffboardingInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Offboarding" }
  /offboardings/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      operationId: updateOffboarding
      tags: [offboarding]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OffboardingUpdate" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Offboarding" }
components:
  schemas:
...
    ApprovalStep:
      type: object
      required: [name, status]
      properties:
        name: { type: string, description: "Unit Head | Department Head | Auto" }
        status: { type: string, description: "pending | approved | rejected" }
        actor: { type: string, nullable: true }
        note: { type: string, nullable: true }
        timestamp: { type: string, nullable: true }
    ApprovalAction:
      type: object
      required: [decision]
      properties:
        decision: { type: string, description: "approve | reject" }
        actor: { type: string, nullable: true }
        note: { type: string, nullable: true }
    LeaveRequest:
      type: object
      required: [id, employeeId, leaveType, startDate, endDate, days, reason, status, currentStep, steps, createdAt]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        leaveType: { type: string, description: "VL | SL" }
        startDate: { type: string }
        endDate: { type: string }
        days: { type: number }
        reason: { type: string }
        status: { type: string, description: "pending | approved | rejected" }
        currentStep: { type: string }
        steps:
          type: array
          items: { $ref: "#/components/schemas/ApprovalStep" }
        createdAt: { type: string }
    LeaveInput:
      type: object
      required: [employeeId, leaveType, startDate, endDate, days, reason]
      properties:
        employeeId: { type: integer }
        leaveType: { type: string }
        startDate: { type: string }
        endDate: { type: string }
        days: { type: number }
        reason: { type: string }
    LeaveBalance:
      type: object
      required: [employeeId, vlBalance, slBalance, vlUsed, slUsed]
      properties:
        employeeId: { type: integer }
        vlBalance: { type: number }
        slBalance: { type: number }
        vlUsed: { type: number }
        slUsed: { type: number }
    RequestType:
      type: string
      enum: [overtime, loan, certificate, reliever, training]
    HrRequest:
      type: object
      required: [id, employeeId, type, title, details, status, currentStep, steps, createdAt]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        type: { $ref: "#/components/schemas/RequestType" }
        title: { type: string }
        details: { type: string }
        status: { type: string }
        currentStep: { type: string }
        steps:
          type: array
          items: { $ref: "#/components/schemas/ApprovalStep" }
        createdAt: { type: string }
    HrRequestInput:
      type: object
      required: [employeeId, type, title, details]
      properties:
        employeeId: { type: integer }
        type: { $ref: "#/components/schemas/RequestType" }
        title: { type: string }
        details: { type: string }
...
    Offboarding:
      type: object
      required: [id, employeeId, reason, exitInterview, hrCleared, itCleared, financeCleared, status, replacementJobId, createdAt]
      properties:
        id: { type: integer }
        employeeId: { type: integer }
        reason: { type: string }
        exitInterview: { type: string }
        hrCleared: { type: boolean }
        itCleared: { type: boolean }
        financeCleared: { type: boolean }
        status: { type: string, description: "Pending | Completed" }
        replacementJobId: { type: integer, nullable: true }
        createdAt: { type: string }
    OffboardingInput:
      type: object
      required: [employeeId, reason]
      properties:
        employeeId: { type: integer }
        reason: { type: string }
        exitInterview: { type: string, nullable: true }
        createReplacementJob: { type: boolean, nullable: true }
    OffboardingUpdate:
      type: object
      properties:
        exitInterview: { type: string, nullable: true }
        hrCleared: { type: boolean, nullable: true }
        itCleared: { type: boolean, nullable: true }
        financeCleared: { type: boolean, nullable: true }
        status: { type: string, nullable: true }